BUG: Fix pixels colored 255 were not segmented property
Because MATLAB is 1 indexed, you can't use an uint8 to index a
256 length array because the zero index must be bumped to 1
and the 255 index must be bumped to 256, but it cant because
that overflows.

+++ tomography.m
@@ -317,7 +317,7 @@ classdef tomography
                 for chunk_start = 1:stride:z
                     chunk = stack(:,:,chunk_start:min([chunk_start+stride-1;z]));
                     % Color Segmentation
-                    chunk = obj.labels{key}(chunk + 1);
+                    chunk = obj.labels{key}(uint16(chunk) + 1);
                     chunk = removeislands(chunk, 8, 100);
                     stack(:,:,chunk_start:min([chunk_start+stride-1;z])) = chunk;
                 end
